% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBU_master_report.R
\name{auto_group_reports}
\alias{auto_group_reports}
\title{Automatically group reports based on naming conventions}
\usage{
auto_group_reports(reports)
}
\arguments{
\item{reports}{List of report metadata}
}
\value{
Named list of grouped reports
}
\description{
Automatically group reports based on naming conventions
}
\keyword{internal}
