% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBU_master_report.R
\name{custom_group_reports}
\alias{custom_group_reports}
\title{Group reports based on user-defined categories}
\usage{
custom_group_reports(reports, report_groups)
}
\arguments{
\item{reports}{List of report metadata}

\item{report_groups}{Named list of patterns for each group}
}
\value{
Named list of grouped reports
}
\description{
Group reports based on user-defined categories
}
\keyword{internal}
