% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBU_master_report.R
\name{extract_report_metadata}
\alias{extract_report_metadata}
\title{Extract metadata from an HTML report file}
\usage{
extract_report_metadata(html_path, metadata)
}
\arguments{
\item{html_path}{Path to the HTML file}

\item{metadata}{List with default metadata values}
}
\value{
Updated metadata list
}
\description{
Extract metadata from an HTML report file
}
\keyword{internal}
