% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_report_style.R
\name{report_custom_transformers}
\alias{report_custom_transformers}
\title{Custom Styler Function}
\usage{
report_custom_transformers()
}
\value{
A named list of styler transformers
}
\description{
Custom Styler Function
}
