% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/underlineHtml.R
\name{underlineHtml}
\alias{underlineHtml}
\title{when selecting a text elements, it add tags to underline it in an HTML report}
\usage{
underlineHtml()
}
\description{
when selecting a text elements, it add tags to underline it in an HTML report
}
