% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBU_create_new_project.R
\name{MBU_create_new_project}
\alias{MBU_create_new_project}
\title{Create the custom project structure}
\usage{
MBU_create_new_project(path, project_id, project_name, ...)
}
\arguments{
\item{path}{The path to the newly created project directory}

\item{project_id}{The project unique identifier}

\item{project_name}{Name of the project}

\item{...}{Additional arguments passed by RStudio (not usually needed)}
}
\value{
Invisible TRUE
}
\description{
This function is called automatically by RStudio after the user chooses the
"My Custom Project" template.
}
