% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBU_new_analysis.R
\name{MBU_new_analysis}
\alias{MBU_new_analysis}
\title{Create Rmd files and folder structure needed, in the current project folder.}
\usage{
MBU_new_analysis(file_name)
}
\arguments{
\item{filename}{name of the new analysis file, it will be used also as title.}

\item{...}{Additional arguments passed by RStudio (not usually needed)}
}
\value{
Invisible TRUE
}
\description{
This function is called automatically by RStudio after the user chooses the
"My Custom Project" template.
}
