% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapInQuote.R
\name{wrapInQuote}
\alias{wrapInQuote}
\title{when selecting a vector of unquoted elements, it add quotes to each element}
\usage{
wrapInQuote()
}
\description{
when selecting a vector of unquoted elements, it add quotes to each element
}
